add no encryption logic
