+++ crates/external_services/src/no_encryption/core.rs
@@ -1,0 +1,17 @@
+//! No encryption core functionalities
+
+/// No encryption type
+#[derive(Debug, Clone)]
+pub struct NoEncryption;
+
+impl NoEncryption {
+    /// Encryption functionality
+    pub fn encrypt(&self, data: impl AsRef<[u8]>) -> Vec<u8> {
+        data.as_ref().into()
+    }
+
+    /// Decryption functionality
+    pub fn decrypt(&self, data: impl AsRef<[u8]>) -> Vec<u8> {
+        data.as_ref().into()
+    }
+}
